Modified   assets/builtin/materials/BuiltinNoTexture.t3b Modified   assets/builtin/program/BuiltinConstant.t3b Modified   source/step09/Core/CMakeLists.txt Modified   source/step09/Core/Include/Kernel/T3DAgent.h Missing    source/step09/Core/Include/Parser/T3DScriptParser.h Missing    source/step09/Core/Include/Parser/T3DScriptParserBase.h Missing    source/step09/Core/Include/Parser/T3DScriptParserGPUProgram.h Missing    source/step09/Core/Include/Parser/T3DScriptParserMaterial.h Missing    source/step09/Core/Include/Parser/T3DScriptParserPass.h Missing    source/step09/Core/Include/Parser/T3DScriptParserSampler.h Missing    source/step09/Core/Include/Parser/T3DScriptParserTechnique.h Missing    source/step09/Core/Include/Parser/T3DScriptParserTextureUnit.h Modified   source/step09/Core/Include/Resource/T3DGPUConstBuffer.h Modified   source/step09/Core/Include/Resource/T3DGPUConstBufferManager.h Modified   source/step09/Core/Include/Resource/T3DMaterial.h Modified   source/step09/Core/Include/T3DPrerequisites.h Modified   source/step09/Core/Include/T3DTypedef.h Modified   source/step09/Core/Source/Kernel/T3DAgent.cpp Modified   source/step09/Core/Source/Kernel/T3DPass.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParser.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserBase.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserGPUProgram.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserMaterial.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserPass.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserSampler.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserTechnique.cpp Missing    source/step09/Core/Source/Parser/T3DScriptParserTextureUnit.cpp Modified   source/step09/Core/Source/Resource/T3DGPUConstBuffer.cpp Modified   source/step09/Core/Source/Resource/T3DGPUConstBufferManager.cpp Modified   source/step09/Core/Source/Resource/T3DGPUProgramManager.cpp Modified   source/step09/Core/Source/Resource/T3DMaterial.cpp Modified   source/step09/Core/Source/Serializer/T3DBinMaterialReader.cpp Modified   source/step09/Core/protobuf/MaterialScriptObject.pb.cc Modified   source/step09/Core/protobuf/MaterialScriptObject.pb.h Modified   source/step09/Tools/ScriptCompiler/Include/T3DScriptTranslatorNew.h Modified   source/step09/Tools/ScriptCompiler/Include/T3DScriptType.h Modified   source/step09/Tools/ScriptCompiler/Source/T3DScriptTranslatorNew.cpp Modified   source/step09/Tools/ScriptCompiler/protobuf/MaterialScriptObject.pb.cc Modified   source/step09/Tools/ScriptCompiler/protobuf/MaterialScriptObject.pb.h Modified   tools/protobuf/proto/MaterialScriptObject.proto Modified   tools/protobuf/source/MaterialScriptObject.pb.cc Modified   tools/protobuf/source/MaterialScriptObject.pb.h

+++ source/step09/Tools/ScriptCompiler/Include/T3DScriptType.h
@@ -897,7 +897,7 @@ namespace Tiny3D
     {
         BuiltinConstantType type;
         String              name;
-        uint8_t             elementCount;
+        uint32_t            elementCount;
         BuiltinType         elementType;
         BuiltinType         extraType;
 
@@ -908,7 +908,7 @@ namespace Tiny3D
             , extraType(BT_NONE)
         {}
 
-        BuiltinConstantDefinition(BuiltinConstantType t, const String &n, uint8_t c, BuiltinType elementT, BuiltinType extraT)
+        BuiltinConstantDefinition(BuiltinConstantType t, const String &n, uint32_t c, BuiltinType elementT, BuiltinType extraT)
             : type(t)
             , name(n)
             , elementCount(c)
